Optionally avoid transmitting sensitive system information
This adds the dontIdentifyToClients option. When set to true, try to avoid
transmitting local host name, port, and time zone. When Polipo is
used with Tor, this information might be useful to an adversary
who wants to identify an anonymous user.

This issue was raised by Gregory Fleischer in his DC 17 presentation,
"Attacking Tor at the Application Layer".

http://pseudo-flaw.net/resource/defcon/attach:/defcon-17-fleischer.pdf
http://pseudo-flaw.net/content/defcon/dc-17-demos/

+++ Non-author changes +++

This change combines the following commits into a single commit :-

https://github.com/jech/polipo/commit/c78beb81e506683a2dc29cfa2eae3c438652585e
https://github.com/jech/polipo/commit/b77065f150546a29a92ac468dbaed88aaf53a9b4
https://github.com/jech/polipo/commit/be116b561321cddc1e1a474e6cefcbf58b7b5e44

Minor trailing whitespace cleanup.